+++ app/Http/Controllers/RoleController.php
@@ -16,8 +16,13 @@ class RoleController extends Controller
      */
     public function index()
     {
-        $roles = Role::all();
-        return view('roles.index', ['roles' => $roles]);
+        if (Auth::user()->can('view-permissions')) {
+            $roles = Role::all();
+            return view('roles.index', ['roles' => $roles]);
+        }
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();        
     }
 
     /**
@@ -26,7 +31,7 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    {        
         //
     }
 
@@ -38,36 +43,29 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        $role =  new Role;
-
-        $this->validate($request, [
-            'name' => 'required',
-            'description' => 'required',
-        ]);
-
-        $role->name = $request->input('name');
-        $role->description = $request->input('description');
-
-        try {
-            $role->save();
-            Session::flash('success', 'Role created Successfully.');
-            return redirect()->back();
-        } catch (\Throwable $th) {
-            Session::flash('error', 'Sorry, something went wrong.');
-            return redirect()->back();
+        if (Auth::user()->can('administer-permissions')) {
+            $role =  new Role;
+    
+            $this->validate($request, [
+                'name' => 'required',
+                'description' => 'required',
+            ]);
+    
+            $role->name = $request->input('name');
+            $role->description = $request->input('description');
+    
+            try {
+                $role->save();
+                Session::flash('success', 'Role created Successfully.');
+                return redirect()->back();
+            } catch (\Throwable $th) {
+                Session::flash('error', 'Sorry, something went wrong.');
+                return redirect()->back();
+            }
         }
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        $role = Role::findOrFail($id);
-        return view('roles.show')->with('role', $role);
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -78,8 +76,13 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findOrFail($id);
-        return view('roles.edit')->with('role', $role);
+        if (Auth::user()->can('administer-permissions')) {
+            $role = Role::findOrFail($id);
+            return view('roles.edit')->with('role', $role);
+        }
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -91,25 +94,29 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $role = Role::findOrFail($id);
+        if (Auth::user()->can('administer-permissions')) {
+            $role = Role::findOrFail($id);
 
-        $this->validate($request, [
-            'name' => 'required',
-            'description' => 'required',
-        ]);
+            $this->validate($request, [
+                'name' => 'required',
+                'description' => 'required',
+            ]);
 
-        $role->name = $request->input('name');
-        $role->staff_id = $request->input('staff_id');
+            $role->name = $request->input('name');
+            $role->staff_id = $request->input('staff_id');
 
-        try {
-            $role->save();
-            Session::flash('success', 'Role updated Successfully.');
-            return redirect()->back();
-        } catch (\Throwable $th) {
-            Session::flash('error', 'Sorry, something went wrong.');
-            return redirect()->back();
+            try {
+                $role->save();
+                Session::flash('success', 'Role updated Successfully.');
+                return redirect()->back();
+            } catch (\Throwable $th) {
+                Session::flash('error', 'Sorry, something went wrong.');
+                return redirect()->back();
+            }
         }
-
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -120,15 +127,18 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
-        
-        try {
-            Role::destroy($id);
-            Session::flash('success', 'Role deleted Successfully.');
-            return redirect()->back();
-        } catch (\Throwable $th) {
-            Session::flash('error', 'Sorry, something went wrong.');
-            return redirect()->back();
+        if (Auth::user()->can('administer-permissions')) {
+            try {
+                Role::destroy($id);
+                Session::flash('success', 'Role deleted Successfully.');
+                return redirect()->back();
+            } catch (\Throwable $th) {
+                Session::flash('error', 'Sorry, something went wrong.');
+                return redirect()->back();
+            }
         }
-
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 }
