+++ app/Http/Controllers/UserController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Session;
+use Auth;
+use Hash;
 
 class UserController extends Controller
 {
+    
     /**
      * Display a listing of the resource.
      *
@@ -15,7 +18,13 @@ class UserController extends Controller
      */
     public function index()
     {
-        return view('users.index');
+        if (Auth::user()->can('view-users')) {
+            $users = User::all();
+            return view('users.index', ['users'=>$users]);
+        }
+        // abort(403);
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
     
     /**
@@ -25,7 +34,12 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('users.create');
+        if (Auth::user()->can('add-users')) {
+            return view('users.create');
+        }
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -36,31 +50,35 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $user = new User;
-        
-        $this->validate($request, [
-            'firstname' => ['required', 'string', 'max:255'],
-            'lastname' => ['required', 'string', 'max:255'],
-            'username' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
-            'password' => ['required', 'string', 'min:6', 'confirmed'],
-        ]);
-
-        $user->firstname = $request->input('firstname');
-        $user->lastname = $request->input('lastname');
-        $user->username = $request->input('username');
-        $user->email = $request->input('email');
-        $user->password = Hash::make($request->input('password'));
-
-        try {
-            $user->save();
-            Session::flash('success', 'User created Successfully.');
-            return redirect()->back();            
-        } catch (\Throwable $th) {
-            Session::flash('error', 'Sorry, something went wrong.');
-            return redirect()->back();
+        if (Auth::user()->can('add-users')) {
+            $user = new User;
+            
+            $this->validate($request, [
+                'firstname' => ['required', 'string', 'max:255'],
+                'lastname' => ['required', 'string', 'max:255'],
+                'username' => ['required', 'string', 'max:255'],
+                'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
+                'password' => ['required', 'string', 'min:6', 'confirmed'],
+            ]);
+    
+            $user->firstname = $request->input('firstname');
+            $user->lastname = $request->input('lastname');
+            $user->username = $request->input('username');
+            $user->email = $request->input('email');
+            $user->password = Hash::make($request->input('password'));
+    
+            try {
+                $user->save();
+                Session::flash('success', 'User created Successfully.');
+                return redirect()->back();            
+            } catch (\Throwable $th) {
+                Session::flash('error', 'Sorry, something went wrong.');
+                return redirect()->back();
+            }
         }
-
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -71,8 +89,13 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        $user = User::findOrFail($id);
-        return view('users.show', ['user' => $user]);
+        if (Auth::user()->can('view-users')) {
+            $user = User::findOrFail($id);
+            return view('users.show', ['user' => $user]);
+        }
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -83,8 +106,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = User::findOrFail($id);
-        return view('users.edit', ['user' => $user]);
+        if (Auth::user()->can('edit-users')) {
+            $user = User::findOrFail($id);
+            return view('users.edit', ['user' => $user]);
+        }
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();
     }
 
     /**
@@ -96,37 +124,41 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $user = User::findOrFail($id);
-
-        $this->validate($request, [
-            'firstname' => ['required', 'string', 'max:255'],
-            'lastname' => ['required', 'string', 'max:255'],
-            'username' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255'],
-        ]);
-        
-        $user->firstname = $request->input('firstname');
-        $user->lastname = $request->input('lastname');
-        $user->username = $request->input('username');
-        $user->email = $request->input('email');
-        
-        if($request->input('password')){
+        if (Auth::user()->can('edit-users')) {
+            $user = User::findOrFail($id);
+    
             $this->validate($request, [
-                'password' => ['required', 'string', 'min:6', 'confirmed'],
+                'firstname' => ['required', 'string', 'max:255'],
+                'lastname' => ['required', 'string', 'max:255'],
+                'username' => ['required', 'string', 'max:255'],
+                'email' => ['required', 'string', 'email', 'max:255'],
             ]);
             
-            $user->password = Hash::make($request->input('password'));
-        }
-
-        try {
-            $user->save();
-            Session::flash('success', 'User updated Successfully.');
-            return redirect()->back();            
-        } catch (\Throwable $th) {
-            Session::flash('error', 'Sorry, something went wrong.');
-            return redirect()->back();
+            $user->firstname = $request->input('firstname');
+            $user->lastname = $request->input('lastname');
+            $user->username = $request->input('username');
+            $user->email = $request->input('email');
+            
+            if($request->input('password')){
+                $this->validate($request, [
+                    'password' => ['required', 'string', 'min:6', 'confirmed'],
+                ]);
+                
+                $user->password = Hash::make($request->input('password'));
+            }
+    
+            try {
+                $user->save();
+                Session::flash('success', 'User updated Successfully.');
+                return redirect()->back();            
+            } catch (\Throwable $th) {
+                Session::flash('error', 'Sorry, something went wrong.');
+                return redirect()->back();
+            }
         }
         
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();        
     }
 
     /**
@@ -137,14 +169,18 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        
-        try {
-            User::destroy($id);
-            Session::flash('success', 'User deleted Successfully.');
-            return redirect()->back();            
-        } catch (\Throwable $th) {
-            Session::flash('error', 'Sorry, something went wrong.');
-            return redirect()->back();
+        if (Auth::user()->can('delete-users')) {
+            try {
+                User::destroy($id);
+                Session::flash('success', 'User deleted Successfully.');
+                return redirect()->back();            
+            } catch (\Throwable $th) {
+                Session::flash('error', 'Sorry, something went wrong.');
+                return redirect()->back();
+            }
         }
+        
+        Session::flash('error', 'Sorry, you are not authorized to access this resource.');
+        return redirect()->back();       
     }
 }
